platform: ae350: support SRST extension
This patch adds early init stage to check whether SMU and WDT are
available in FDT. If so, their addresses will be stored in smu_data
and wdt_data, and the ae350_reset will be added to system reset device
list to support cold reboot and shutdown types. Note that we cannot
poweroff the board with SMU, the shutdown will simply hang in M-mode.

Signed-off-by: Yu Chien Peter Lin <peterlin@andestech.com>
Reviewed-by: Dylan Jhong <dylan@andestech.com>