ae350/plicsw: Modify IPI scheme on PLIC-SW
Current IPI scheme uses bit 0 of Pending Reigster on PLIC-SW
to send IPI from hart 0 to hart 7, but bit 0 needs to be hardwired to 0 according to spec.
Therefore, to support 8 core bitmaps, we have to re-order the IPI scheme.

After some investigation, self-IPI seems to be seldom or never used,
so we have bit 0 indicate sending IPI from hart 0 to hart 0.
Please refer to the comment for more detailed description.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>