IPI-CMO: Bypass 4KB-size check in RFENCE framework
IPI-CMO implementation is based on RFENCE framework
where 4KB-size check will be asserted in sbi_tlb_update.

We bypass this check for it is only for RFENCE framework,
we should be able to do cache operation for size that is greater than 4KB.

We borrow asid member of struct sbi_tlb_info to distinguish
IPI-CMO from RFENCE. The maxium ASID length for RV64 is 16 bit,
so we could set it to a value larger than the maxium value ASID
can represent. Currently we use 0x53434D4F, the ASCII of SCMO(SBI_CMO),
to decide if it is IPI-CMO that we are processing.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>