ae350: fix light sleep hang issue
The sbi_printf in light sleep flow seems to
cause an immediate escape from first wfi that
should be executed in opensbi and then enter
the wfi in kernel. Therefore, light sleep hart
will hang in kernel wfi and could not be woken
up.

Fix this by removing the sbi_printf.
Tested in the order of thousands of times.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>