platform: generic: andes: Moving PMA mapping table to per hart scratch region
In the past, we used to store the PMA mapping table information in a
single global array, which is actually incorrect. In Andes' design,
PMA belongs to the CSR of each hart. Using a single array to share
PMA table among each hart is not correct. Therefore, we moved the
PMA mapping table to the sbi_scratch region. Sbi_scratch is designed
in OpenSBI to store individual information for each hart.

Placing the PMA mapping table in sbi_scratch is a better design than
putting it in a single array.

Fixes: ed478347793e ("platform: generic: andes: Add PMA SBI call")
Signed-off-by: Dylan Jhong <dylan@andestech.com>