platform: generic: andes: Remove the error message in mcall_free_pma()
If a memory region wants to change its memory attribute from cacheable
to non-cacheable in Linux kernel, it needs to use ioremap_wc() to call
mcall_set_pma() to set PMA. And OpenSBI will record the VA in the PMA
mapping table. However, for a device region, since its default attribute
is non-cacheable, it usually does not need to set PMA to change its memory
attribute. Therefore, the kernel will call ioremap() but not mcall_set_pma().

Currently, both memory and device regions will call iounmap() in the kernel
and this function will eventually call mcall_free_pma(). When the driver
(device region) calls iounmap(), it will fail to find the corresponding VA
in the PMA mapping table.

It is more complicated to check whether to create PA in memory or device in
OpenSBI, as OpenSBI is not easy to convert from VA to PA. Therefore, we
ultimately decided to adopt a simpler method of removing the error message
to solve this problem.

Fixes: ed478347793e ("platform: generic: andes: Add PMA SBI call")
Signed-off-by: Dylan Jhong <dylan@andestech.com>