lib: ipi: Rearrange Andes PLICSW to single-bit-per-hart strategy
Source hart infomation is not necessary in IPI, so we could
use single-bit-per-hart strategy to rearrange PLICSW mapping.

Bit 0 of Interrupt Pending Bits is hardwired to 0.
Therefore, we use bit 1 to send IPI to hart 0,
bit 2 to hart 1, ..., and so on.

Andes AE350 platform is guaranteed to have 31 interrupt sources.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>