lib: sbi: Introduce sbi_printf_highlight() API
This patch introduces a new API sbi_printf_highlight(). It is used to
highlight the printing messages in OpenSBI. Currently, we only highlight
the error messages for PMA. If there are other error messages we want to
highlight in the future, this API can help us easily spot the messages.

Signed-off-by: Ben Zong-You Xie <ben717@andestech.com>