platform: generic: andes: Add the "print_detailed_cause" method
When multiple events cause a trap to be taken with the same mcause value,
we need to know the mdcause value that records the detailed cause.
The "print_detailed_cause" method is utilised to print out the detailed
cause when the unhandled trap occurs. In addition, we also pass the
mdcause value to sdcause while redirecting the trap.

Reformed from the following patches on ast-v5_3_0-branch
- (4bae0611ad) platform: generic: andes: Fix an error for print_detailed_cause() on imprecise exception (#85)

Signed-off-by: Ben Zong-You Xie <ben717@andestech.com>