andes/sleep.S: fix system hang caused by hotplugging cpu (#89)
Different sleep modes run through different code path, and we use
ae350_suspend_mode to determine which sleep mode the hart should perform.

We found that we may retrieve the wrong value if accessing ae350_suspend_mode
after disabling D cache. The reason is that we only write-back-invalidate D cache
but not both D cache and L2 cache. Therefore, the updated value only exists in L2,
which is now invisible to the hart for that D cache is disabled.

Accessing ae350_suspend_mode would now be loading a stale value from RAM,
hotplug command would execute code that should not be executed, causing
the system to hang.

Fix this by storing the value of ae350_suspend_mode in $s1,
and do not perform another load after disabling D cache.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/opensbi/pulls/89
Reviewed-by: Peter Yu-Chien Lin <peterlin@andestech.com>
Reviewed-by: Dylan Dai-Rong Jhong <dylan@andestech.com>
Reviewed-by: Tim Shih-Ting OuYang <tim609@andestech.com>